Rename Coder.Desktop.Rpc to Coder.Desktop.Vpn
