+++ Tests/Vpn.Proto/RpcHeaderTest.cs
@@ -1,7 +1,7 @@
 using System.Text;
-using Coder.Desktop.Rpc.Proto;
+using Coder.Desktop.Vpn.Proto;
 
-namespace Coder.Desktop.Tests.Rpc.Proto;
+namespace Coder.Desktop.Tests.Vpn.Proto;
 
 [TestFixture]
 public class RpcHeaderTest
