+++ Tests/Vpn/SerdesTest.cs
@@ -1,9 +1,9 @@
 using System.Buffers.Binary;
-using Coder.Desktop.Rpc;
-using Coder.Desktop.Rpc.Proto;
+using Coder.Desktop.Vpn;
+using Coder.Desktop.Vpn.Proto;
 using Google.Protobuf;
 
-namespace Coder.Desktop.Tests.Rpc;
+namespace Coder.Desktop.Tests.Vpn;
 
 [TestFixture]
 public class SerdesTest
