+++ Vpn.Proto/Vpn.Proto.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <RootNamespace>Coder.Desktop.Rpc.Proto</RootNamespace>
+        <RootNamespace>Coder.Desktop.Vpn.Proto</RootNamespace>
         <TargetFramework>net8.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
