+++ Vpn.Proto/vpn.proto
@@ -1,6 +1,6 @@
 ﻿syntax = "proto3";
 option go_package = "github.com/coder/coder/v2/vpn";
-option csharp_namespace = "Coder.Desktop.Rpc.Proto";
+option csharp_namespace = "Coder.Desktop.Vpn.Proto";
 
 import "google/protobuf/timestamp.proto";
